Langchain/requirements.txt
langchain
python-dotenv
langchain_community
wikipedia
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit